+++ src/common-typeorm/entities/index.ts
@@ -1,0 +1,17 @@
+// Discord
+export { DiscordGuildEntity } from './discord/discord-guild.entity';
+export { DiscordUserEntity } from './discord/discord-user.entity';
+
+// Event
+export { EventEntity } from './event/event.entity';
+export { EventDiscordDetailsEntity } from './event/event-discord-details.entity';
+export { EventDiscordParticipantEntity } from './event/event-discord-participant.entity';
+
+// POAP
+export { PoapLinkEntity } from './poap/poap-link.entity';
+export { PoapLinkDiscordDetailsEntity } from './poap/poap-link-discord-details.entity';
+
+// Next Auth
+export { AccountEntity } from './auth/account.entity';
+export { UserEntity } from './auth/user.entity';
+export { TokenEntity } from './auth/token.entity';