+++ src/index.ts
@@ -21,9 +21,11 @@ export { CommonPinoLoggerModule } from './common-pino-logger/common-pino-logger.
 export { CommonPinoLoggerService } from './common-pino-logger/common-pino-logger.service';
 
 // Config
-export { joiValidationConfig } from './config/joi-validation.config';
-export { configureCacheOptions } from './config/redis.config';
-export { configureBullOptions } from './config/redis.config';
+export { CommonConfigModule } from './common-config/common-config.module';
+export { RedisConfigModule } from './redis-config/redis-config.module';
+export { RedisBullConfigModule } from './redis-bull-config/redis-bull-config.module';
+export { DiscordConfigModule } from './discord-config/discord-config.module';
+export { MongooseConfigModule } from './mongoose-config/mongoose-config.module';
 
 // Utils
 export { parseReleaseUtil } from './utils/parse-release.util';
@@ -33,3 +35,8 @@ export { loadRedisCacheUtil } from './utils/load-redis-cache.util';
 // DTOs
 export { CommunityEventDto } from './dto/redis/community-event.dto';
 export { DiscordParticipantDto } from './dto/redis/discord-participant.dto';
+
+// Entities - TypeORM
+export { CommonTypeOrmModule } from './common-typeorm/common-type-orm.module';
+export * as entities from './common-typeorm/entities';
+export * from './common-typeorm/entities';